refactor retrieve_model moving in slaid.utils
